+++ js-api/src/interfaces/d4.ts
@@ -1927,6 +1927,12 @@ export interface IGridSettings {
 
   marginBottom: number;
 
+  /// Heatmap horizontal scroll positions (minValue, maxValue)
+  heatmapHorzScroll: Array<number>;
+
+  /// Heatmap vertical scroll positions (minValue, maxValue)
+  heatmapVertScroll: Array<number>;
+
   /// Determines whether newly added columns are added to the grid
   syncNewColumns: boolean;